Address review follow-up issues

+++ internal/imap/safe.go
@@ -9,7 +9,7 @@ func safeMessageCount(count int) uint32 {
 	if count < 0 {
 		return 0
 	}
-	if count > math.MaxUint32 {
+	if uint64(count) > uint64(math.MaxUint32) {
 		log.Printf("IMAP: clamping message count %d to uint32 max", count)
 		return math.MaxUint32
 	}
